Changes to make this code functional
Linked the actual configuration values to the parameters to posixmq.open()
Wrapped posixmq.open() in try/catch so a failure to open would result in a red status on the node.
Removed some debug lines
Queue is not unlinked by default on node removal (TODO: Consider adding this in as a config option)
Add a respective posixmq-write node to send messages to a queue
Removed dependency on input to probe for pending messages. Messages are emitted as they appear in the queue.
Removed input from PosixMQ-read node; it only has an output.